% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogasoutcomesmalawi.R
\docType{data}
\name{biogasoutcomesmalawi}
\alias{biogasoutcomesmalawi}
\title{Qualitative interviews about the outcomes of biogas reactors in Malawi}
\format{
A tibble with 5963 rows and 5 variables:
\describe{
\item{interview_id}{ The site id where the interview was performed.}
\item{interview_date}{ The date on which the interview was performed.}
\item{question_id}{ The number of the question within each interview. Note: Number of questions differ between interviews.}
\item{interviewer}{ The question asked by the interviewer.}
\item{interviewee}{ The answer given by the interviewee.}
}
}
\usage{
biogasoutcomesmalawi
}
\description{
This dataset consists of 61 semi-structured interviews with biogas
owners in the Southern Region of Malawi. Interviews were conducted over
multiple visits to sites over the course of a two-year period between
June 2021 and October 2022.
}
\keyword{datasets}
